SOFT FORK: Today, BitClout joins the environmental movement. (#26)
I am proud to announce that the dev community has committed to moving to
a zero-waste consensus mechanism, and that it is taking a major step in
that direction with a new update.

Cryptocurrencies have gotten a lot of heat recently for being harmful to
the environment.

This is due in large part to "proof of work," a consensus mechanism that
burns CPU power to secure the network.

BitClout has relied on proof of work to get off the ground, but since
launching the dev community has already invested in consensus mechanisms
that allow BitClout to operate on orders of magnitude less power than
most existing proof of work cryptocurrencies.

Now, in a show of solidarity with the environmental movement, and to
take the first big step toward a zero-waste consensus mechanism,
BitClout is deploying a new update that reduces its power consumption by
a full order of magnitude.

BitClout is reducing the block reward, and thus the amount of
environmental waste its consensus mechanism incurs, by a factor of ten.

This was a difficult change to make, and it took the input of several
major node operators to deploy it safely.

We understand that miners in particular may not like this change;
however, there is now consensus around the decision to push the limits
of the BitClout chain toward conservation rather than allowing the
depletion of what are ultimately finite environmental resources.

There is still more work to do to move to a zero-waste consensus
mechanism without compromising on decentralization (e.g. full proof of
stake).

But the dev community is committed to getting there sooner rather than
later, and we hope this change shows just how committed we are.