Fix bug introduced by NFT splits
What was happening?
* Before, the royalty UTXO indices would start at len(txn.TxOutputs), as they should.
* NFT splits introduced a bug where they start at len(txn.TxOutputs)+1
on accident.

This bug is virtually impossible to find unless you sync a chain from
scratch, which I did as a final test.

This is because txn assembly will
happily use the incremented UTXOs rather than the unincremented ones and
it won't actually break anything in a test environment as long as the
txn construction code is the same as the connect() code.