Fix stubborn NFT splits bug due to reordering of royalties in map
This was a fun one. Because NFT splits were being deduped using a map,
the royalties would wind up producing UTXOs in a nondeterministic order.
This would cause two nodes to have a different value for the UTXO at a
particular index.

The fix was to sort the royalties deterministically before paying them
out. Easy-peasy.