Fix DESO balance check in DAO coin limit order
The way we were validating DESO balance made it so that we could only
execute a "sell" of DESO if the resulting balance was above the sell
amount. This made it so that the maximum amount of DESO we could sell in
an order was ~(DESO balance / 2), rather than (DESO - fee).

The fix was to move the execution of _connectBasicTransfer to *after*
the point where all the validations were done. This had no impact other
than fixing our validation logic.