Mf/add limit order metadata order type field (#342)
* Add limit order metadata OrderType field.

* Update postgres moar.

* Dynamically determine order book transactor fees.

* Add helper function for calculating FeeNanos.

* Use uint8s for enums.

* Add txn metadata validations.

* Small refactor as vars used only once.

* Add ImmediateOrCancel order type.

* Allow zero price on IOC and FOK orders.

* Allow user to specify exchange rate on IOC and FOK orders.

* Add tests for market ASK orders.

* Store OrderEntry.OrderType in the database.

* diamondhands touchups

* Require OrderType must be specified. No default value.

* Refactor IsMarketOrder util function location.

* Validate txn metadata.

* Rename OrderType to FillType.

* Update test description.

* Fix failing CI.

* Fix DESO balance check in DAO coin limit order

The way we were validating DESO balance made it so that we could only
execute a "sell" of DESO if the resulting balance was above the sell
amount. This made it so that the maximum amount of DESO we could sell in
an order was ~(DESO balance / 2), rather than (DESO - fee).

The fix was to move the execution of _connectBasicTransfer to *after*
the point where all the validations were done. This had no impact other
than fixing our validation logic.

* Catch an error when parsing fill type

* Add more uint8 overflow checks.

Co-authored-by: diamondhands <diamondhands@bitcloutdev.com>