Speed up blocksync by only writing to db when an entry has changed
In order to optimize flushes for certain prefixes, we want to only write entries to the db that have changed.
On top of that, we add a further optimization to only update related indexes if the stake amount or status has changed in the
validator. Not doing this results in a lot of writes to badger
every epoch that eventually slow block processing to a crawl. This is essentially a bug in
badger when you repeatedly write to the same key, and we're papering over it here in response
to encountering the issue. In an ideal world, badger would work as intended and this extra
optimization wouldn't be necessary.

Prefixes affected:
* PrefixStakeByValidatorAndStaker
* PrefixStakeByStakeAmount
* PrefixValidatorByPKID
* PrefixValidatorByStatusAndStakeAmount
* PrefixSnapshotValidatorSetByPKID
* PrefixSnapshotValidatorSetByStakeAmount